Remove redundant results screen and simplify game flow
- Remove ResultsScreen.tsx and associated routing
- Update game flow to transition directly from voting to post-game
- Add confetti animation to PostGameScreen for winners
- Update monitor-game-timers to handle voting → completed transition
- Fix database transition_game_status function for new flow
- Update all routing logic and phase transitions
- Update test scripts to work with simplified flow

Game flow is now: Briefing → Drawing → Voting → Post-Game (completed)
